requirements.txt
python-dotenv>=1.0
flask>=2.0
requests>=2.28
numpy>=1.24
pandas>=2.0
geopy>=2.2
google-generativeai>=0.3.0
